+++ tests/testthat/test-graph.R
@@ -19,19 +19,6 @@ test_with_dir("Supplied graph is not an igraph.", {
   expect_error(prune_drake_graph(12345, to = "node"))
 })
 
-test_with_dir("visNetwork dep graph does not fail if input file is binary", {
-  skip_on_cran() # CRAN gets whitelist tests only (check time limits).
-  skip_if_not_installed("visNetwork")
-  x <- drake_plan(
-    y = readRDS(file_in("input.rds")),
-    strings_in_dots = "literals"
-  )
-  saveRDS(as.list(mtcars), "input.rds")
-  con <- drake_config(x, verbose = FALSE)
-  expect_silent(out <- vis_drake_graph(con))
-  unlink("input.rds", force = TRUE)
-})
-
 test_with_dir("null graph", {
   skip_on_cran() # CRAN gets whitelist tests only (check time limits).
   x <- drake_graph_info(config = list(graph = igraph::make_empty_graph()))
@@ -77,20 +64,6 @@ test_with_dir("build_drake_graph() works", {
   config <- dbug()
   expect_equal(
     class(build_drake_graph(config$plan, verbose = FALSE)), "igraph")
-})
-
-test_with_dir("dependency visNetwork runs", {
-  skip_on_cran()
-  skip_if_not_installed("visNetwork")
-  config <- dbug()
-  pdf(NULL)
-  tmp <- vis_drake_graph(config)
-  dev.off()
-  pdf(NULL)
-  tmp <- vis_drake_graph(config, full_legend = FALSE)
-  dev.off()
-  unlink("Rplots.pdf", force = TRUE)
-  expect_true(is.character(default_graph_title()))
 })
 
 test_with_dir("Supplied graph is pruned.", {
@@ -157,12 +130,6 @@ test_with_dir("we can generate different visNetwork dependency graphs", {
   expect_equal(sort(outdated(config = config)),
                sort(c(config$plan$target)))
   expect_false(file.exists("Makefile"))
-
-  file <- "graph.html"
-  expect_false(file.exists(file))
-  skip_if_not_installed("visNetwork")
-  vis_drake_graph(config = config, file = file)
-  expect_true(file.exists(file))
 })
 
 test_with_dir("clusters", {
@@ -231,19 +198,6 @@ test_with_dir("can get the graph info when a file is missing", {
   expect_true("missing" %in% o$nodes$status)
 })
 
-test_with_dir("static graphs", {
-  skip_on_cran()
-  skip_if_not_installed("ggplot2")
-  skip_if_not_installed("ggraph")
-  load_mtcars_example()
-  config <- drake_config(
-    my_plan, cache = storr::storr_environment(), session_info = FALSE)
-  gg <- static_drake_graph(config)
-  expect_true(inherits(gg, "ggplot"))
-  make(config = config)
-  gg <- static_drake_graph(config)
-  expect_true(inherits(gg, "ggplot"))
-  if ("package:ggraph" %in% search()){
-    eval(parse(text = "detach('package:ggraph', unload = TRUE)"))
-  }
+test_with_dir("misc graph utils", {
+  expect_true(is.character(default_graph_title()))
 })